+++ incidenttemplate.go
@@ -60,6 +60,7 @@ func (c *Client) CreateIncidentTemplate(input *CreateIncidentTemplateInput) (*Cr
 type GetIncidentTemplatesInput struct {
 	_ struct{}
 	// an integer specifying the starting point (beginning with 0) when paging through a list of entities
+	// Default: 0
 	StartIndex *int
 
 	// the maximum number of results when paging through a list of entities.
@@ -82,9 +83,13 @@ func (c *Client) GetIncidentTemplates(input *GetIncidentTemplatesInput) (*GetInc
 	q := url.Values{}
 	if input.StartIndex != nil {
 		q.Add("start-index", strconv.Itoa(*input.StartIndex))
+	} else {
+		q.Add("start-index", "0")
 	}
 	if input.MaxResults != nil {
 		q.Add("max-results", strconv.Itoa(*input.MaxResults))
+	} else {
+		q.Add("max-results", "50")
 	}
 
 	resp, err := c.httpClient.R().Get(fmt.Sprintf("%s?%s", apiRoutes.incidentTemplates, q.Encode()))