+++ statuspage.go
@@ -89,6 +89,7 @@ func (c *Client) CreateStatusPage(input *CreateStatusPageInput) (*CreateStatusPa
 type GetStatusPagesInput struct {
 	_ struct{}
 	// an integer specifying the starting point (beginning with 0) when paging through a list of entities
+	// Default: 0
 	StartIndex *int
 
 	// the maximum number of results when paging through a list of entities.
@@ -114,9 +115,13 @@ func (c *Client) GetStatusPages(input *GetStatusPagesInput) (*GetStatusPagesOutp
 	q := url.Values{}
 	if input.StartIndex != nil {
 		q.Add("start-index", strconv.Itoa(*input.StartIndex))
+	} else {
+		q.Add("start-index", "0")
 	}
 	if input.MaxResults != nil {
 		q.Add("max-results", strconv.Itoa(*input.MaxResults))
+	} else {
+		q.Add("max-results", "50")
 	}
 
 	for _, include := range input.Include {